Made vitter_algorithms header-only
